Added more drawing options and line tooltips
drawDots, drawLegend and addTooltips can now be passed in as options to
a line chart. Also added addLineTooltips option that will display
interpolated data when hovering over a line. (The goal was to make it
possible to display lots of data in a graph by enabling line tooltips
and disabling dots)