Make y2 axis shown only when y2 is used.
Hi! I was getting the y2 axis displayed in situations where it was not being specified in any series, and the y parameter of the axes object is being modified.  I believe this is a bug and not the intent, but if it is please disregard this.  If you take a look at the "y0 - y1" example (http://n3-charts.github.io/line-chart/#/examples) you will see that the y2 axis is shown, but no series targets it.  If "axes: y {something}" is defined, then you will get both y and y2 being drawn.  This change corrects this.  Thanks!